Central_Postgre/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.12.5
flet==0.28.3
requests==2.32.5
twilio==9.10.1
python-dotenv==1.1.1
pysnmp==7.1.22
pycryptodomex
psycopg2-binary
